+++ Zolian.Server.Base/Network/Components/PlayerRegenerationComponent.cs
@@ -88,7 +88,7 @@ public class PlayerRegenerationComponent(WorldServer server) : WorldServerCompon
 
         try
         {
-            currentHp += (int)Math.Clamp(seed, 5, cap);
+            currentHp += (long)Math.Clamp(seed, 5, cap);
             client.Aisling.CurrentHp = currentHp;
 
             if (client.Aisling.CurrentHp > client.Aisling.MaximumHp)
@@ -107,7 +107,7 @@ public class PlayerRegenerationComponent(WorldServer server) : WorldServerCompon
 
         try
         {
-            currentMp += (int)Math.Clamp(seed, 5, cap);
+            currentMp += (long)Math.Clamp(seed, 5, cap);
             client.Aisling.CurrentMp = currentMp;
 
             if (client.Aisling.CurrentMp > client.Aisling.MaximumMp)
